split out non-private nested classes
